rename prompter to `image_prompter`
