+++ test/contrib/test_prompter.py
@@ -3,7 +3,7 @@ import sys
 import pytest
 import torch
 
-from kornia.contrib.prompter import ImagePrompter
+from kornia.contrib.image_prompter import ImagePrompter
 from kornia.contrib.sam import SamConfig
 from kornia.testing import BaseTester
